Move error action consts into error controller to inline enums
