+++ src/controller/content-steering-controller.ts
@@ -2,7 +2,7 @@ import { Events } from '../events';
 import { Level } from '../types/level';
 import { AttrList } from '../utils/attr-list';
 import { addGroupId } from './level-controller';
-import { ErrorActionFlags, NetworkErrorAction } from '../errors';
+import { ErrorActionFlags, NetworkErrorAction } from './error-controller';
 import { logger } from '../utils/logger';
 import type Hls from '../hls';
 import type { NetworkComponentAPI } from '../types/component-api';
