+++ src/errors.ts
@@ -1,6 +1,3 @@
-import { RetryConfig } from './config';
-import { HdcpLevel } from './types/level';
-
 export enum ErrorTypes {
   // Identifier for a network error (loading error / timeout ...)
   NETWORK_ERROR = 'networkError',
@@ -89,29 +86,3 @@ export enum ErrorDetails {
   // Uncategorized error
   UNKNOWN = 'unknown',
 }
-
-export const enum NetworkErrorAction {
-  DoNothing = 0,
-  SendEndCallback = 1, // Reserved for future use
-  SendAlternateToPenaltyBox = 2,
-  RemoveAlternatePermanently = 3, // Reserved for future use
-  InsertDiscontinuity = 4, // Reserved for future use
-  RetryRequest = 5,
-}
-
-export const enum ErrorActionFlags {
-  None = 0,
-  MoveAllAlternatesMatchingHost = 1,
-  MoveAllAlternatesMatchingHDCP = 1 << 1,
-  SwitchToSDR = 1 << 2, // Reserved for future use
-}
-
-export type IErrorAction = {
-  action: NetworkErrorAction;
-  flags: ErrorActionFlags;
-  retryCount?: number;
-  retryConfig?: RetryConfig;
-  hdcpLevel?: HdcpLevel;
-  nextAutoLevel?: number;
-  resolved?: boolean;
-};
